+++ include/perfcpp/analyzer/memory_access.hpp
@@ -8,6 +8,7 @@
 #include <perfcpp/sample/result.hpp>
 #include <set>
 #include <string>
+#include <string_view>
 #include <typeinfo>
 #include <unordered_map>
 #include <unordered_set>
@@ -34,15 +35,9 @@ public:
 
   [[nodiscard]] std::string to_string() const;
   [[nodiscard]] std::string to_json() const;
-  [[nodiscard]] std::string to_csv(const std::string& data_type_name,
+  [[nodiscard]] std::string to_csv(std::string_view data_type_name,
                                    char delimiter = ',',
                                    bool print_header = true) const;
-  [[nodiscard]] std::string to_csv(std::string&& data_type_name,
-                                   const char delimiter = ',',
-                                   const bool print_header = true) const
-  {
-    return to_csv(data_type_name, delimiter, print_header);
-  }
 
 private:
   std::vector<DataType> _data_types;