+++ include/perfcpp/sample/result.hpp
@@ -3,6 +3,7 @@
 #include <perfcpp/sample/recording_values.hpp>
 #include <fstream>
 #include <functional>
+#include <string_view>
 #include <type_traits>
 #include <utility>
 #include <vector>
@@ -151,14 +152,14 @@ public:
    * @param delimiter Character used to separate CSV columns (default: ',').
    * @param list_delimiter Character used to separate list elements within a cell (default: ';').
    */
-  void to_csv(std::string&& file_name, char delimiter = ',', char list_delimiter = ';') const;
+  void to_csv(std::string_view file_name, char delimiter = ',', char list_delimiter = ';') const;
 
   /**
    * Writes the sample results as flamegraphs input to the given file.
    *
    * @param file_name File to write the sample results in a format that can be read by flame graph generators.
    */
-  void to_flamegraphs(std::string&& file_name) const;
+  void to_flamegraphs(std::string_view file_name) const;
 
 private:
   /// List of values recorded by the sample. These values are represented in the samples.